Tanks v0.6.1 - Added tutorial, improved the level generator to make more appealing levels with newer obstacles, improved tank AI, tweaked motion, changed the firework animation, changed how your levels are deleted - there is now a delete button, and empty levels will no longer be deleted automatically. Also includes other improvements.

+++ src/lwjglwindow/FontRenderer.java
@@ -59,9 +59,9 @@ public class FontRenderer
 		}
 	}
 	
-	public int getStringSizeX(double sX, String s)
+	public double getStringSizeX(double sX, String s)
 	{
-		int w = 0;
+		double w = 0;
 		char[] c = s.toCharArray();
 		
 		for (int i = 0; i < c.length; i++)
@@ -72,8 +72,8 @@ public class FontRenderer
 		return w;
 	}
 	
-	public int getStringSizeY(double sY, String s)
+	public double getStringSizeY(double sY, String s)
 	{
-		return (int) (sY * 32);
+		return (sY * 32);
 	}
 }
